+++ tests/Models/TestProduct.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Cyclechain\Sandboxer\Tests\Models;
+
+use Illuminate\Database\Eloquent\Model;
+
+class TestProduct extends Model
+{
+    protected $table = 'products';
+    protected $guarded = [];
+}