+++ tests/Models/TestUser.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Cyclechain\Sandboxer\Tests\Models;
+
+use Illuminate\Foundation\Auth\User as Authenticatable;
+
+class TestUser extends Authenticatable
+{
+    protected $table = 'users';
+    protected $guarded = [];
+}